backend/requirements.txt
fastapi
uvicorn
pillow
transformers
torch
python-dotenv
huggingface_hub
